+++ src/main/java/qna/service/QnaService.java
@@ -11,6 +11,8 @@ import qna.domain.*;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import qna.repository.AnswerRepository;
+import qna.repository.QuestionRepository;
 
 @Service
 public class QnaService {
