+++ src/test/java/qna/domain/DeleteHistoryTest.java
@@ -7,7 +7,7 @@ import static qna.domain.UserTest.SANJIGI;
 
 import java.time.LocalDateTime;
 
-class DeleteHistoryTest {
+public class DeleteHistoryTest {
     public static final DeleteHistory D1 = new DeleteHistory(ContentType.ANSWER, A1.getId(), JAVAJIGI.getId(),
             LocalDateTime.now());
     public static final DeleteHistory D2 = new DeleteHistory(ContentType.ANSWER, A2.getId(), SANJIGI.getId(),
