+++ src/test/java/qna/repository/AnswerRepositoryTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -10,6 +10,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.NotFoundException;
+import qna.domain.Answer;
+import qna.domain.AnswerTest;
+import qna.domain.QuestionTest;
+import qna.domain.UserTest;
 
 @DataJpaTest
 class AnswerRepositoryTest {
