+++ src/test/java/qna/repository/DeleteHistoryRepositoryTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.DeleteHistory;
+import qna.domain.DeleteHistoryTest;
 
 @DataJpaTest
 class DeleteHistoryRepositoryTest {
