+++ src/test/java/qna/repository/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.NotFoundException;
+import qna.domain.User;
+import qna.domain.UserTest;
 
 @DataJpaTest
 class UserRepositoryTest {
